refactor: move the message transformations to its own file
